+++ ChangeLog.md
@@ -1,6 +1,8 @@
 ## change log ##
 
 * 2017-02-24:  3.2.4
+  - SelfHostedOntController: adjustments in resolution to dispatch request for 
+    organization or user
   - GET /org/{orgName} now with optional withOnts=yes to include the list of
     registered ontologies by the given organization.